Revert backref support until Source/Delta refactor
The backref code was merged before refactor to ensure that the
abstractions chosen for the refactor would be powerful enough to support
backref demands. I also wanted to avoid a painful rebase.

This turned out to be a bad idea, so, reverting.

Revert "backref: factor out into own file"

This reverts commit 93fd00235f85fd89e74b4961c5f028b478b7964b.

Revert "Add backref support (#17)"

This reverts commit 7c8208c890f71c6cad186afc3dc089b89caa71a7.